feat: productIds selector 구현
