+++ src/components/pages/ProductsPage/ProductList/ProductList.tsx
@@ -1,4 +1,9 @@
+import { useCallback, useEffect } from 'react';
+import { useRecoilValue } from 'recoil';
+
 import { Product } from '@customTypes/Product';
+import { productIdsSelector } from '@recoil/selector';
+import useFetch from '@hooks/useFetch';
 
 import { StyledProductList } from '@components/pages/ProductsPage/ProductList/ProductList.styled';
 import ProductItem from '@components/pages/ProductsPage/ProductList/ProductItem/ProductItem';
@@ -6,26 +11,20 @@ import ErrorModal from '@components/pages/ErrorPage/ErrorModal/ErrorModal';
 import { FetchedDataList } from '@components/commons/FetchedDataList/FetchedDataList';
 
 const ProductList = () => {
-  // TODO : 페이지 벗어남 (새로고침, 뒤로가기, 페이지 이동) 감지시 데이터 패칭
-  // const { getData, postData, patchData, deleteData } = useFetch('/cart-items');
+  const { postData, patchData, deleteData } = useFetch('/cart-items');
+  const productIds = useRecoilValue(productIdsSelector);
 
-  // const addItemOnCart = useCallback(async () => {
-  //   try {
-  //     await postData({ productId: productId });
-  //   } catch (error) {
-  //     console.log(error);
-  //   }
-  // }, [postData, productId]);
+  const handleProductListPageExit = useCallback(() => {
+    window.addEventListener('beforeunload', () => {
+      // postData();
+    });
+  }, []);
 
-  // const deleteItemOnCart = useCallback(async () => {
-  //   try {
-  //     const cartItems = await getData<Product[]>();
-  //     const item = cartItems.find(item => item.id === productId);
-  //     await deleteData(`/${item?.id}`);
-  //   } catch (error) {
-  //     console.log(error);
-  //   }
-  // }, [getData, deleteData, productId]);
+  useEffect(() => {
+    handleProductListPageExit();
+  }, [handleProductListPageExit]);
+
+  console.log(productIds);
 
   return (
     <FetchedDataList<Product[]> endpoint={'/products'} initialValue={[]}>
