+++ src/recoil/atom.ts
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+import { CartItems } from '@customTypes/Product';
+
+export const cartItemsState = atom<CartItems>({
+  key: 'cartItemsState',
+  default: {},
+});
